==> validation/combined_FischerEtAl/02_spillover_correction_com.Rmd <==
---
title: "02_spillover_matrix"
author: "Lasse Meyer"
date: "1/24/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Spillover correction

based on the [CATALYST](https://bioconductor.org/packages/release/bioc/html/CATALYST.html) R package. 
For the original publication, see [here](https://www.cell.com/cell-systems/fulltext/S2405-4712(18)30063-2).

## Software requirements

```{r lib, message=FALSE,warning=FALSE}
#Data containers
library(SpatialExperiment)
library(SingleCellExperiment)

#Data processing
library(CATALYST)
library(scuttle)
library(scater)
library(imcRtools)

#Data visualization
library(cytomapper)
library(dittoSeq)
library(pheatmap)

#Utilities
library(tidyverse)
library(openxlsx)
library(stringr)
library(patchwork)
library(BiocParallel)
library(tiff)
```

## Overview
1. Reading in the data
2. Quality control 
3. (Optional) pixel binning
4. "Debarcoding" for pixel assignment
5. Pixel selection for spillover matrix estimation
6. Spillover matrix generation
7. Saving the results
8. Single-cell compensation
9. Image compensation

## Generate the spillover matrix

Generated once for each panel used and saved for later use. 

### Read in the data

Individual .txt files into a `SingleCellExperiment` object (can directly be used by the `CATALYST` package). The .txt file names need to contain the spotted metal isotope name (format `(mt)(mass)` - like `Sm152`) 

For further downstream analysis, we will asinh-transform the data using a
cofactor of 5.

```{r read-txts-transform, message=FALSE}
sce_spill <- readSCEfromTXT("/mnt/rcc_volume/TNBC/spillover/spillover_renamed/")

#Transformation
assay(sce_spill, "exprs") <- asinh(counts(sce_spill)/5)
```

### Quality control

In the next step, we will observe the median pixel intensities per spot and
threshold on medians < 200 counts.
These types of visualization serve two purposes:

1. Small median pixel intensities (< 200 counts) might hinder the robust
estimation of the channel spillover. In that case, consecutive pixels can be
summed (see [Optional pixel binning](#pixel_binning)).

2. Each spotted metal (row) should show the highest median pixel intensity in its
corresponding channel (column). If this is not the case, either the naming of the
.txt files was incorrect or the incorrect metal was spotted.

```{r QC-heatmap, message = FALSE, fig.width=7, fig.height=7}
# Log10 median pixel counts per spot and channel
plotSpotHeatmap(sce_spill)

# Thresholded on 200 pixel counts
plotSpotHeatmap(sce_spill, log = FALSE, threshold = 200)
```

**Observations:** 
Wrong labeling of ROIs (renaming was performed)
No signal = Er167, Eu153, In113, Sm147
Completely wrong signal = Er168 (Dy162 spotted twice)


### Optional pixel binning {#pixel_binning}

In cases where median pixel intensities are low (< 200 counts), which is the case here, consecutive
pixels can be summed to increase the robustness of the spillover estimation.
The `imcRtools` package provides the `binAcrossPixels` function,
which performs aggregation for each channel across `bin_size` consecutive pixels 
per spotted metal.

```{r binning, message=FALSE, fig.width=7, fig.height=7}
# Define grouping
bin_size = 5

sce_spill_2 <- binAcrossPixels(sce_spill, bin_size = bin_size)

assay(sce_spill_2, "exprs") <- asinh(counts(sce_spill_2)/5)

# Log10 median pixel counts per spot and channel
plotSpotHeatmap(sce_spill_2)

# Thresholded on 200 pixel counts
plotSpotHeatmap(sce_spill_2, log = FALSE, threshold = 200)
```

### Filtering incorrectly assigned pixels

Based on the intensity distribution of all channel, pixels are assigned to their corresponding barcode; here this is the already known metal spot. This procedure serves the purpose to identify pixels that cannot be robustly assigned to the spotted metal. Pixels of such kind can be regarded as "noisy", "background" or "artefacts" that should be removed prior to spillover estimation.

The general workflow for pixel debarcoding is as follows:
1. assign a preliminary metal mass to each pixel
2. for each pixel, estimate a cutoff parameter for the distance between 
positive and negative pixel sets
3. apply the estimated cutoffs to identify truly positive pixels

```{r debarcoding, message=FALSE}
bc_key <- as.numeric(unique(sce_spill$sample_mass))
bc_key <- bc_key[order(bc_key)]

#Preliminary metal mass for pixel
sce_spill_2 <- assignPrelim(sce_spill_2, bc_key = bc_key)

#estimate and apply cut-offs
sce_spill_2 <- estCutoffs(sce_spill_2)
sce_spill_2 <- applyCutoffs(sce_spill_2)
```

The obtained `SingleCellExperiment` now contains the additional `bc_id` entry.
For each pixel, this vector indicates the assigned mass (e.g. `161`) or
`0`, meaning unassigned. 

This information can be visualized in form of a heatmap:

```{r assignment-heatmap, fig.width=7, fig.height=7}
cur_table <- table(sce_spill_2$bc_id, sce_spill_2$sample_mass)

pheatmap::pheatmap(log10(cur_table + 1),
         cluster_rows = FALSE,
         cluster_cols = FALSE)
```

However, in cases where incorrect assignment occurred or where few pixels were
measured for some spots, the `imcRtools` package exports a simple helper
function to exclude pixels based on these criteria:

```{r pixel-filtering}
sce_spill_2 <- filterPixels(sce_spill_2, minevents = 40, correct_pixels = TRUE)
```

### Compute spillover matrix

Based on the single-positive pixels, we use the `CATALYST::computeSpillmat()`
function to compute the spillover matrix and `CATALYST::plotSpillmat()` to
visualize it. The `plotSpillmat` function checks the spotted and acquired
metal isotopes against a pre-defined `CATALYST::isotope_list()`. In this data,
the `Ar80` channel was additionally acquired to check for deviations in signal
intensity. `Ar80` needs to be added to a custom `isotope_list` object for
visualization.

Here was also add an external spillover matrix and compare the two.

```{r compute-spillover, fig.width=7, fig.height=7, warning=FALSE}
sce_spill_2 <- computeSpillmat(sce_spill_2)
sm <- as.data.frame(metadata(sce_spill_2)$spillover_matrix)

isotope_list <- CATALYST::isotope_list
isotope_list$Ar <- 80

#Load Janas spillover matrix for the same panel (aquired at similiar time-points and the same panel)
sm_jana <- read.csv("/mnt/rcc_volume/TNBC/data_analysis/03_data/spillover_matrix_Jana.csv",row.names = 1)
metadata(sce_spill_2)$sm_jana <- sm_jana

#Plot both spillover matrices
plotSpillmat(sce_spill_2,sm_jana, isotope_list = isotope_list) #Janas
plotSpillmat(sce_spill_2, isotope_list = isotope_list) #My - with pixel binning

#Check that spillover matrices contain are metals that are included in the study
panel <- read.csv("/mnt/rcc_volume/TNBC/panel.csv")
channels <- paste0(panel$channel[panel$keep == 1], "Di")

channels[which(!channels %in% rownames(sm))]
channels[which(!channels %in% rownames(sm_jana))]

#Add missing Ho165 to Dy164/Er166 spillover estimates from my experiment to Janas spillover matrix (used for downstream analysis)
sm_jana <- rbind(sm_jana,sm[rownames(sm) == "Ho165Di",])
```

**Observations:**
- Very high oxidative spillover between 139 and 155 (>10% in my experiment)
- Not clearly visible on the pictures (some bright spots with high ion counts may be responsible - see ZTMA249_ID_56_AX8Y7, ZTMA249_ID_31_AX7Y4)

### Save spillover matrix as csv

Next, we will write out the spillover matrix as .csv file for possible later use.

```{r write_sm_csv}
#sm
#write.csv(sm, "/mnt/rcc_volume/TNBC/data_analysis/03_data/spillover_matrix_LM_mod.csv")

sm_jana
write.csv(sm, "/mnt/tnbc_volume/TNBC_all/data_analysis/03_data/spillover_matrix_Jana_mod.csv")
```


## Single-cell data compensation

Firstly, we will use the `CATALYST` package to perform spillover compensation on the
**single-cell mean intensities** using the `compCytof`function.
Load `SpatialExperiment` object generated in previous script.

```{r single-cell-compensation}
sce <- readRDS("/mnt/tnbc_volume/TNBC_all/data_analysis/01_sce_TNBC_com.rds")

rowData(sce)$channel_name <- paste0(rowData(sce)$channel, "Di")

# #sce <- compCytof(sce, sm, 
#                  transform = TRUE, cofactor = 1,
#                  isotope_list = isotope_list, 
#                  overwrite = FALSE)
sce_jana <- compCytof(sce, sm_jana, 
                 transform = TRUE, cofactor = 1,
                 isotope_list = isotope_list, 
                 overwrite = FALSE)
```

To check the effect of channel spillover compensation, the expression of markers
that are affected by spillover can be visualized in form of scatter plots before and after
compensation. Here we focus on oxidative spillover between La139(H3K27me3) and Gd155(Bcl2)

```{r visualize-single-cell-spillover, message=FALSE}
#Simple scatterplot
before <- dittoScatterPlot(sce, x.var = "H3K27me3", y.var = "Bcl2",
                           assay.x = "exprs", assay.y = "exprs") +
    ggtitle("Before compensation")

# after <- dittoScatterPlot(sce, x.var = "H3K27me3", y.var = "Bcl2",
#                           assay.x = "compexprs", assay.y = "compexprs") +
#     ggtitle("After compensation")

after_1 <- dittoScatterPlot(sce_jana, x.var = "H3K27me3", y.var = "Bcl2",
                          assay.x = "compexprs", assay.y = "compexprs") +
    ggtitle("After compensation - Jana")

before+after_1

#Smoothed density scatterplot
buylrd = c("#313695", "#4575B4", "#74ADD1", "#ABD9E9", "#E0F3F8", "#FFFFBF",
           "#FEE090", "#FDAE61", "#F46D43", "#D73027", "#A50026") 
myColRamp = colorRampPalette(c(buylrd))


smoothScatter(x=assay(sce,"exprs")[rownames(sce) == "H3K27me3"],
              y=assay(sce,"exprs")[rownames(sce) == "Bcl2"],
              colramp=myColRamp,
              main="Before compensation",
              xlab="H3K27me3 (La139)",
              ylab="Bcl2 (Gd155)")

# smoothScatter(x=assay(sce,"compexprs")[rownames(sce) == "H3K27me3"],
#               y=assay(sce,"compexprs")[rownames(sce) == "Bcl2"],
#               colramp=myColRamp,
#               main="After compensation",
#               xlab="H3K27me3 (La139)",
#               ylab="Bcl2 (Gd155)")
  

smoothScatter(x=assay(sce_jana,"compexprs")[rownames(sce_jana) == "H3K27me3"],
              y=assay(sce_jana,"compexprs")[rownames(sce_jana) == "Bcl2"],
              colramp=myColRamp,
              main="After compensation - Jana",
              xlab="H3K27me3 (La139)",
              ylab="Bcl2 (Gd155)")
```

**We will use the spillover matrix from Jana for downstream analysis.** 

To facilitate further downstream analysis, the non-compensated assays can now be
replaced by their compensated counterparts:

```{r overwrite-assays}
assay(sce, "counts") <- assay(sce_jana, "compcounts") 
assay(sce, "exprs") <- assay(sce_jana, "compexprs") 
assay(sce, "compcounts") <- assay(sce, "compexprs") <- NULL
```

## Image compensation

Secondly, we will use the [cytomapper](https://github.com/BodenmillerGroup/cytomapper) package to perform spillover compensation on the
**multi-channel images** using the `compImage`function.
Load `CytoImageList` object generated in previous script.

```{r read-in-image, message=FALSE}
images <- readRDS("/mnt/rcc_volume/TNBC/data_analysis/images_segtest.rds")
channelNames(images) <- rowData(sce)$channel_name
```

Create an adapted spillover matrix that matches the `channelNames` of the
`CytoImageList` object and can be used to perform pixel-level spillover
compensation.

```{r image-compensation, message = FALSE}
panel <- read.csv("/mnt/rcc_volume/TNBC/panel.csv")
sm_jana <- as.matrix(sm_jana)
adapted_sm <- adaptSpillmat(sm_jana, paste0(panel$channel[panel$keep == 1], "Di"),isotope_list = isotope_list)

images_comp <- compImage(images, adapted_sm, overwrite = FALSE,BPPARAM = MulticoreParam(workers = 14,progressbar = TRUE,stop.on.error = FALSE))
```

**Issue: For some reason the images stored on disk are overwritten.**
As a sanity check, we will visualize the image before and after compensation. 

```{r image-visualization}
#still to be performed - see issue above
```

For convenience, we will re-set the `channelNames` to their biological targtes:

```{r re-set-channels}
channelNames(images_comp) <- rownames(sce)
```

## Save objects

For further downstream analysis, the compensated `SpatialExperiment` and
`CytoImageList` objects are saved:

```{r save, message = F}
saveRDS(sce, "/mnt/tnbc_volume/TNBC_all/data_analysis/02_sce_TNBC_spillcor_com.rds")
#saveRDS(images_comp, "/mnt/rcc_volume/TNBC/data_analysis/images_TNBC_spillcor.rds")
```
